Switch domain:
https://github.com/creek-service/creek-test/issues/24
