+++ example/src/test/java/org/creekservice/internal/example/ExampleImplTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Creek Contributors (https://github.com/creek-service)
+ * Copyright 2021-2022 Creek Contributors (https://github.com/creek-service)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,20 @@
  * limitations under the License.
  */
 
-package org.creek;
+package org.creekservice.internal.example;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
 
-import org.creekservice.api.test.conformity.ConformityTester;
+import org.creekservice.api.example.Example;
 import org.junit.jupiter.api.Test;
 
-class ModuleTest {
+class ExampleImplTest {
+
+    private final Example example = new ExampleImpl();
 
     @Test
-    void shouldConform() {
-        ConformityTester.test(ModuleTest.class);
+    void shouldReturnTrue() {
+        assertThat(example.getTrue(), is(true));
     }
 }
